[WEAV-220] 상태 변경 검증 및 finished 조건 변경
