+++ domain/src/main/kotlin/com/studentcenter/weave/domain/meeting/enums/MeetingStatus.kt
@@ -2,9 +2,10 @@ package com.studentcenter.weave.domain.meeting.enums
 
 enum class MeetingStatus {
     PENDING,
-    PREPARING,
-    SCHEDULED,
-    INSUFFICIENT,
+    // FIXME: 현재 사용하지 않는 상태(MVP 스펙 X)
+    // PREPARING,
+    // SCHEDULED,
+    // INSUFFICIENT,
     COMPLETED,
     CANCELED;
 }
